requirements.txt
ibm-db==3.1.0
ibm-db-sa==0.3.7
SQLAlchemy==1.4.27
pandas==1.3.4
gunicorn==20.1.0
Flask==2.0.2
Flask-Cors==3.0.10
gevent
python-dotenv==0.19.2
numpy==1.21.4
scikit-learn==1.0.1
xgboost==1.5.1
